DatabaseConnectionListener is now only available to gemini-jdbc
