+++ gemini/src/main/java/com/techempower/gemini/GeminiApplication.java
@@ -140,7 +140,6 @@ public abstract class GeminiApplication
   private final ImageHelper                imageHelper;
   private final EntityStore                entityStore;
   private final GeminiLocaleManager        localeManager;
-  private final DatabaseConnectionListener dbConnListener;
   private final EntityUpdater              entityUpdater;
   private final GeminiMonitor              monitor;
   private final PyxisSecurity              security;
@@ -214,7 +213,6 @@ public abstract class GeminiApplication
       this.sessionManager       = constructSessionManager();
       this.simSessionManager    = constructSimSessionManager();
       this.localeManager        = constructLocaleManager();
-      this.dbConnListener       = constructDatabaseConnectionListener();
       this.entityUpdater        = constructEntityUpdater();
   
       final InitDisplayBanner banner = new InitDisplayBanner(); 
@@ -223,9 +221,6 @@ public abstract class GeminiApplication
       
       // Add the start-up notification feature.
       getFeatureManager().add(FEATURE_STARTUP_NOTE, "Send application start-up notification", true);
-      
-      // Give the ConnectorFactory a reference to the database listener.
-      connectorFactory.setDatabaseConnectionListener(dbConnListener);
 
       // Add some configurables that do not add themselves.
       configurator.addConfigurable(connectorFactory);
@@ -658,23 +653,6 @@ public abstract class GeminiApplication
    * Constructs a GeminiMonitor reference.
    */
   protected abstract GeminiMonitor constructMonitor();
-  
-  /**
-   * Construct a DatabaseConnectionListener.  If this returns non-null,
-   * this listener object will be provided to the ConnectorFactory instance
-   * upon its creation.  By default, a list of two listeners is provided:
-   * 
-   * <ul>
-   * <li>A BasicConnectorListener.</li>
-   * <li>And the GeminiMonitor's listener.</li>
-   * </ul>
-   */
-  protected DatabaseConnectionListener constructDatabaseConnectionListener()
-  {
-    return new DatabaseConnectionListenerList(
-        new BasicConnectorListener(this),
-        getMonitor().getListener());
-  }
 
   /**
    * Construct an ImageHelper. Returns a JvnImageHelper by default but can
@@ -945,15 +923,6 @@ public abstract class GeminiApplication
   public String getAdministratorEmail()
   {
     return administratorEmail;
-  }
-  
-  /**
-   * Gets the DatabaseConnectionListener reference.
-   */
-  @Override
-  public DatabaseConnectionListener getDatabaseConnectionListener()
-  {
-    return dbConnListener;
   }
 
   /**
